# api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  description: "The Monika Cloud Cloud Platform API provides web services for the Monika suite of apps and systems."
  version: "0.1.0"
  title: "Monika Cloud Patform API"
  termsOfService: "https://cloud.monika.com/terms/"
  contact:
    email: "swood@monika.com"
  license:
    name: "Commercial"
    url: "http://cloud.monika.com/licenses/license-1.0.html"
host: "127.0.0.1:9000"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
...
  /case/list:
    x-swagger-router-controller: case.controller
    post:
      tags:
      - "case"
      summary: "Get a list of case, filtered and sorted"
      description: "Get a list of case, filtered and sorted"
      operationId: "listCases"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: body
        name: body
        description: "pararmeters for filtering and sorting the list of cases"
        required: false
        schema:
          type: object
          properties:
            tenantListFilterQuery:
              $ref: "#/definitions/CaseListFilterQuery"
            tenantListSortQuery:
              $ref: "#/definitions/CaseListSortQuery"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/Case"
        400:
          description: "Invalid query data"
  /case/{caseId}/comment:
    x-swagger-router-controller: case.controller0
    post:
      tags:
      - "case"
      summary: "add comment to the case"
      description: "upload comment to the case"
      operationId: "addComment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "caseId"
        in: path
        description: "The id of the case id you would like to have the comment"
        required: true
        type: string
        format: uuid
      - in: body
        name: body
        description: "pararmeters for adding comment to the case"
        required: true
        schema:
          $ref: "#/definitions/Comment"
      responses:
        200:
          description: "successful operation"
          schema:
            type: string
            description: "Comment successfully added"
        400:
          description: "Invalid query data"
  /case/comment:
    x-swagger-router-controller: case.controller        
    put:
      tags:
      - "case"
      summary: "Update an existing comment from cases"
      description: "Updates the comment"
      operationId: "editComment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: body
        name: body
        description: "The comment object to be updated"
        required: true
        schema:
          $ref: "#/definitions/Comment"
      responses:
        200:
          description: "Comment successfuly updated"
          schema:
            type: string
        404:
          description: "Comment not found"
        405:
          description: "Invalid input"
  /case/{caseId}/attachment:
    x-swagger-router-controller: case.controller
    post:
      tags:
      - "case"
      summary: "upload attachment to the case"
      description: "upload attachment to the case"
      operationId: "uploadAttachment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "caseId"
        in: path
        description: "The id of the case id you would like to have the attachment"
        required: true
        type: string
        format: uuid
      - in: body
        name: body
        description: "pararmeters for adding attachment to the case"
        required: true
        schema:
          $ref: "#/definitions/Attachment"
      responses:
        200:
          description: "successful operation"
          schema:
            type: string
            description: "Comment successfully added"
        400:
          description: "Invalid query data"
  /case/attachment/{attachmentId}:
    x-swagger-router-controller: case.controller
    delete:
      tags:
      - "case"
      summary: "Delete the attachment by id"
      description: "Delete the attachment by id. Note, this is a permanent operation which cannot be undone. Make a backup first."
      operationId: "deleteAttachment"
      produces:
      - "application/json"
      parameters:
      - name: "attachmentId"
        in: path
        description: "The id of the attachment you would like to delete"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "successful operation, attachment deleted"
          schema:
            type: string
        404:
          description: "attachment not found"
  /case/{caseId}/processStep:
    x-swagger-router-controller: case.controller
    get:
      tags:
      - "case"
      summary: "process steps of a case"
      description: "process steps of a case"
      operationId: "processStep"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "caseId"
        in: path
        description: "The id of the case id you would like to have the comment"
        required: true
        type: string
        format: uuid
      - in: body
        name: body
        description: "parameters for adding comment to the case"
        required: true
        schema:
          type: object
          properties:
              input:
                type: string
                description: "The data inputed by the user through the step"
              inputType:
                type: integer
                description: "The step number to process"  
      responses:
        200:
          description: "successful operation"
          schema:
            type: object
            properties:
              question:
                type: string
                description: "The process question"
              questionType:
                type: integer
                description: "The type of question"
        400:
          description: "Invalid query data"
  /dashboard/list:
    x-swagger-router-controller: dashboard.controller
    get:
      tags:
      - "dashboard"
      summary: "Get a list of widgets, filtered and sorted"
      description: "Get a list of widgets, filtered and sorted"
      operationId: "listWidget"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/Widget"
        400:
          description: "Invalid query data"
  /dashboard:
    x-swagger-router-controller: dashboard.controller
    post:
      tags:
      - "dashboard"
      summary: "Add a new widget to the dashboard"
      description: "Add a new widget to the dashboard"
      operationId: "addWidget"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: body
        name: body
        description: "The widget object to be added"
        required: true
        schema:
          $ref: "#/definitions/Widget"
      responses:
        200:
          description: "Tenant successfuly added"
          schema:
            type: object
            properties:
              tenantId:
                type: string
                format: uuid
        400:
          description: "The widget data posted was not valid"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Invalid input"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "There is a user already matching that data"
          schema:
            $ref: "#/definitions/Error"
  /dashboard/widgetType/list:
    x-swagger-router-controller: dashboard.controller
    get:
      tags:
      - "dashboard"
      summary: "Get a list of permission types"
      description: "Get a list of permissions types"
      operationId: "listWidgetTypes"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/WidgetType"
        400:
          description: "Invalid query data"
  /dashboard/{widgetId}:
    x-swagger-router-controller: dashboard.controller
    patch:
      tags:
      - "dashboard"
      summary: "Update an existing widget's details in the dashboard"
      description: "Updates the widget details"
      operationId: "editWidget"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "widgetId"
        in: path
        description: "The id of the widget you would like to update"
        required: true
        type: string
        format: uuid
      - name: JsonPatch
        in: body
        required: true
        schema:
          $ref: "#/definitions/PatchRequest"
      responses:
        200:
          description: "Tenant successfuly updated"
          schema:
            type: string
        404:
          description: "Tenant not found"
        405:
          description: "Invalid input"  
    delete:
        tags:
        - "dashboard"
        summary: "Deletes existing widget from the dashboard"
        description: "Deletes existing widget from the dashboard"
        operationId: "deleteWidget"
        produces:
        - "application/json"
        parameters:
        - name: "widgetId"
          in: path
          description: "The id of the widget you would like to delete"
          required: true
          type: string
          format: uuid
        responses:
          200:
            description: "successful operation, widget deleted"
            schema:
              type: string
          404:
            description: "Widget not found"        
definitions:
...
  Error:
    type: object
    properties:
      code:
        type: integer
      message:
        type: string
    required:
    - code
    - message
...
  PatchRequest:
    type: array
    items:
      $ref: "#/definitions/PatchDocument"
  PatchDocument: 
    description: A JSONPatch document as defined by RFC 6902 
    required:
     - "op"
     - "path"
    properties: 
     op: 
      type: string 
      description: The operation to be performed 
      enum:
       - "add"
       - "remove"
       - "replace"
       - "move"
       - "copy"
       - "test"
     path: 
      type: string 
      description: A JSON-Pointer 
     value: 
      type: string 
      description: The value to be used within the operations.
     from: 
      type: string 
      description: A string containing a JSON Pointer value.
...
  Case:
    type: object
    properties:
      id:
        type: string
        format: uuid
      region:
        type: string
        description: "Case Region/Site. e.g VIC Metro - 60 Elizabeth Street"
      location:
        type: string
        description: "The Case location, e.g. Main Kitchen- Freezer"
      description:
        type: string
        description: "The Case description e.g The product temperature is too high at 12:00"
      created:
        type: string
        format: date-time  
      unit:
        type: string
        description: "The unit of the case e.g. oven,freezer"
      alarms:
        type: object
        properties:
          alarmId:
            type: string
            format: uuid
          alarmType:
            type: string
      resolution:
        type: string
        description: "The case resolution, e.g resolved,unresolved"
      asignee: 
        type: object
        properties:
          userId:
            type: string
            format: uuid
      status:
        type: string
        description: "Status"
        enum:
        - "Data gathering"
        - "Investigating"
        - "Closed"
  CaseListSortQuery:
    type: object
    properties:
      sortBy:
        type: string
        description: "Sort By"
        enum:
        - "date created"
        - "case description"
        - "assignee"
        - "status"
        - "resolution"
        - "region"
        - "site"
        - "priority"
      order:
        type: string
        description: "Sort Order"
        enum:
        - "ascending"
        - "descending"
  CaseListFilterQuery:
    type: object
    properties:
      caseId:
        type: string
        format: uuid
      regionId:
        type: string
        format: uuid
      siteId:
        type: string
        format: uuid
      roleId:
        type: string
        format: uuid
      status:
        type: string
        description: "Status"
        enum:
        - "active"
        - "deactivated"
      resolution:
        type: string
        description: "perform a keyword search and only show those results with matches"
      priority:
        type: string
        description: "perform a keyword search and only show those results with matches" 
  Attachment:
    type: object
    properties:
      attachmentId:
        type: string
        format: uuid
      attachment:
        type: string
        format: file
  Comment:
    type: object
    properties:
      commentId:
        type: string
        format: uuid
      comment:
        type: string
        description: "comment"
  Widget:
    type: object
    properties:
      id:
        type: string
        format: uuid
      title:
        type: string
        description: "The title of the widget, e.g. Task Completed to Schedule, Time in Alarm"
      region:
        type: string
        description: "The region. e.g. Worcester College"
      site:
        type: string
        description: "The site e.g. Victoria - ABC Cafe Melbourne"
      time:
        type: boolean
        description: "The boolean which tell if the time will be visible in the widget"
      legend:
        type: boolean
        description: "The boolean which tell if the legend will be visible in the widget"
      color:
        type: string
        description: "The color of the widget"  
      type:
        type: string
        description: "Status"
        enum:
        - "donut"
        - "progress"
        - "health"
        - "icon2"
        - "overview"
        - "map"
        - "gauge"
        - "statistics"
        - "bar"
      data:
        description: Can be anything.
    required:
    - data
    - title
    - region
    - site
    - color
    - type
  WidgetType:
    type: object
    properties:
      id:
        type: string
        format: uuid
      name:
        type: string
      description:
        type: string
      tenantScope:
        type: boolean
        description: "Whether this widget can be used with tenant scope"
      regionScope:
        type: boolean
        description: "Whether this widget can be used with region scope"
      siteScope:
        type: boolean
        description: "Whether this widget can be used with site scope"    
...